Make wait methods available to cucumber
